requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==3.0.1
Django==4.2
django-allauth==0.54.0
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-filter==23.2
django-heroku==0.3.1
django-htmx==1.26.0
django-import-export==4.3.10
django-template-partials==25.2
django-widget-tweaks==1.4.12
Faker==18.6.0
gunicorn==23.0.0
idna==3.4
isort==6.1.0
mypy-extensions==1.0.0
narwhals==2.7.0
numpy==2.3.3
oauthlib==3.2.2
packaging==23.1
pandas==2.3.3
pathspec==0.11.1
platformdirs==3.5.0
plotly==6.3.1
psycopg2==2.9.10
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2025.2
requests==2.29.0
requests-oauthlib==1.3.1
setuptools==80.9.0
six==1.16.0
sqlparse==0.4.4
tablib==3.8.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.11.0
gunicorn